@@ -6,8 +6,7 @@ repository: https://github.com/chulwoo-park/timelines/
 issue_tracker: https://github.com/chulwoo-park/timelines/issues
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: '>=1.17.0'
+  sdk: 3.6.0
 
 dependencies:
   flutter:
