@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:timelines/timelines.dart';
 
@@ -74,7 +73,7 @@ class _ThemePageState extends State<ThemePage> {
                     children: [
                       Text(
                         'TimelineTheme',
-                        style: Theme.of(context).textTheme.headline6,
+                        style: Theme.of(context).textTheme.titleMedium,
                       ),
                       _ThemeDropdown(
                         title: 'Direction',
@@ -140,7 +139,7 @@ class _ThemePageState extends State<ThemePage> {
                     children: [
                       Text(
                         'IndicatorTheme',
-                        style: Theme.of(context).textTheme.headline6,
+                        style: Theme.of(context).textTheme.titleMedium,
                       ),
                       _ThemeDropdown<Color?>(
                         title: 'Color',
@@ -192,7 +191,7 @@ class _ThemePageState extends State<ThemePage> {
                     children: [
                       Text(
                         'ConnectorTheme',
-                        style: Theme.of(context).textTheme.headline6,
+                        style: Theme.of(context).textTheme.titleMedium,
                       ),
                       _ThemeDropdown<Color?>(
                         title: 'Color',
