@@ -83,7 +83,7 @@ class ConnectorThemeData with Diagnosticable {
 
   @override
   int get hashCode {
-    return hashValues(
+    return Object.hash(
       color,
       space,
       thickness,
